fq: revert to float

+++ src/dso_display.cpp
@@ -50,15 +50,14 @@ static const uint16_t *getBackGround(int line)
 static const char *fq2Text(int fq)
 {
     static char buff[16];
-    int  f=fq;
+    float  f=fq;
     const char *suff="";
-#define STEP(x,t)  if(f>x)     {suff=t;f/=x;}else
+#define STEP(x,t)  if(f>x)     {suff=t;f/=(float)x;}else
 
     STEP(1000000,"M")
     STEP(1000,"K")
     {}
-    
-    sprintf(buff,"%3d%s",f,suff);
+    sprintf(buff,"%2.1f%s",f,suff);
     return buff;
 }
 /**
